+++ pages/product-overview/product-overview.js
@@ -1,10 +1,39 @@
-export async function initProductOverviewPage() {
+import { API_URL } from "../../settings";
+export async function initProductOverviewPage(searchTerm) {
 
     // Event listener for the button
     const seeProductsButton = document.getElementById("see-products");
   seeProductsButton.addEventListener("click", fetchAndDisplayProducts);
 //til Simone
-  
+
+
+
+searchBtn.addEventListener('click', async function() {
+try{
+  const res = await fetch(API_URL + "/products/detailed")
+  const data = await res.json()
+
+ // Filter the data based on the search term
+ const filteredDataName = data.filter(product => {
+     // You can customize this condition based on your data structure
+      
+     return product.name.toLowerCase().includes(searchTerm)
+ });
+
+ console.log("Filtered DataName:", filteredDataName);
+
+  const filteredDataCategory = data.filter(product => {
+       
+      return product.category.toLowerCase().includes(searchTerm) + product.name.toLowerCase().includes(searchTerm) + product.description.toLowerCase().includes(searchTerm)
+  });
+ 
+  console.log("Filtered DataCAtegory:", filteredDataCategory);
+
+} catch{
+  console.log("virker ikke")
+}
+}
+) 
 }
